// src/Moryx.Products.Web/src/app/api/models/product-definition-model.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Entry } from '@moryx/ngx-web-framework/entry-editor';
export interface ProductDefinitionModel {
  baseDefinition?: string | null;
  displayName?: string | null;
  name?: string | null;
  properties?: Entry;
}