// src/Moryx.Products.Web/src/app/api/models/recipe-classification-model.ts
/*
 * Copyright (c) 2026 Phoenix Contact GmbH & Co. KG
 * Licensed under the Apache License, Version 2.0
*/

/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export enum RecipeClassificationModel {
  Unset = 'Unset',
  Default = 'Default',
  Alternative = 'Alternative',
  Intermediate = 'Intermediate',
  Part = 'Part'
}